Bind Extra IOExpander to a separate enable pin
Binding it to the same enable pin as the other three doesn't provide
enough power...

It may be necessary to setup a screen and display information without
being connected to the PC. Time will tell

The power drop may also affect the three io expanders as well :/ we
shall see what happens.

I have also stopped using the MCP reset pin as it is not necessary at
all.