For the ft232h interface, circuitpython is the quickest way to implementation
I have not decided whether to use SRAM + Atmega1284p or ft232h
standalone.

Each style introduces extra complexities but the ultimate goal is to use
the ram on my desktop as the memory block for the i960Sx. It really
doesn't matter where the memory comes from, as long as it is provided.