The chipset debug interface is now the chipset tweakables interface
I have an idea for a future improvement to use the dma spi engine in the mcu as a way to zero out the bss section.

The unit would be called the pattern commit engine to the i960. It would take in a start, length, and a pattern to write. Then a doorbell would be written to as the signal. In the first generation, the processor would insert wait states until the pattern engine returned. Future designs will use the HOLD/HLDA pins to talk to the i960 from the mcu. Currently, those pins are tied to an spi expander. Instead I want to directly tie those to the microcontroller's pins. But that also requires some more reworks as well. If I am going to do that then I need to reuse the "spi bus" pins as an 8-bit input port for address and a number of signals from the i960.

This new idea requires a heavy redesign that I am not in the mood to do right now.