Make newDataCycle responsible for querying what updateData cycle would normally do.
Keeping updateDataCycle around for now. It is more expensive with what it is doing since it is being doubled up.

Also, introduced burstNext to increment the current address and query the byte enable bits. At this point, we do not need to keep sampling the burst address bits on each iteration past the first since we already know that burst transactions do not jump around