interrupts are just always on when doing the enhanced port write operations
doing some research online, digitalWrite disables interrupts on each digitalWrite call because the routine could be called from an interrupt. Their goal was to be as simple as possible regardless of performance overhead. Since I don't use interrupts and thus port manipulation during an interrupt it is safe to remove all references to automatically disabling/enabling interrupts in the related functions.

This has the side effect of simplifying the code somewhat without affecting readability.