0xFE00'0000 is now io configuration space
While more wasteful, I am going to spread the different mappings throughout memory starting from 0xFD00'0000 and going down.

It is totally configurable and the monitor rom reads the configuration space on boot to set everything up. So changes I make are not going to be a huge deal on either side of the barrier.