Change the cache configuration to be:
Eight-Way 64-Element Cache with 128-bytes per line

This seems to have the highest hit rate compared to lower way counts or higher byte widths. The increase in hit count means that less time overall is being spent on the psram spi bus. The downside is that each time a miss happens up to 264 bytes will be transferred across the spi bus to and from the psram. The upside to this latency is that there is only at most 8 bytes of overhead from the spi address versus the continual overhead from transferring smaller things