Rearrange full 32-bit update so that we can take advantage of the unused cycles to do some level of computation.
Now we pull the upper 16 first and then check to see if the target functions need to be updated while clocking out the commands to the lower 16 io expander.

Even if the comparison overruns the spi clock out frame it doesn't matter because that the time being used would have been wasted anyways.